restructure app system
