+++ src/pages/index.js
@@ -1,13 +1,15 @@
-import Hero from "./Hero"
-import Program from "./Program"
-import About from "./About"
-import Campus from "./Campus"
-import Testimonials from "./Testimonials"
+import Hero from "./Hero/Hero"
+import Program from "./Program/Program"
+import About from "./About/About"
+import Campus from "./Campus/Campus"
+import Testimonials from "./Testimonials/Testimonials"
+import Contact from "./Contact/Contact"
 
 export {
     Hero,
     Program,
+    Contact,
     About,
     Testimonials,
-    Campus
+    Campus,
 }
